movie_footprints/requirements/base.txt
Django==4.2.4
django-environ==0.10.0
Pillow==10.0.0
psycopg2-binary==2.9.7

djangorestframework==3.14.0
django-rest-knox==4.2.0
